Properly space out section on pricing page.
Previously the pricing page was using section elements and inline
styling to space out individual portions of the page. However, the
section element seems to be intended primarily for large segmented
portions of a page, such as the design for the home page. This and the
inline styling has been replaced with the .pricing-section class.